moved get_conf to utils
