+++ tests/stress/sync_stress/test_sync.py
@@ -2,9 +2,10 @@ from datetime import datetime
 import time
 from pytest_testconfig import config as testconfig
 
-from tests.test_bs import add_multi_clients, get_conf, setup_bootstrap, save_log_on_exit
+from tests.test_bs import add_multi_clients, setup_bootstrap, save_log_on_exit
 import tests.queries as q
 from tests.convenience import convert_ts_to_datetime
+from tests.utils import get_conf
 
 # ==============================================================================
 #    TESTS
@@ -21,7 +22,7 @@ def test_sync_stress(init_session, setup_bootstrap, save_log_on_exit):
 
     clients_num = testconfig["client"]["replicas"]
     bs_info = setup_bootstrap.pods[0]
-    cspec = get_conf(bs_info, testconfig['client'])
+    cspec = get_conf(bs_info, testconfig['client'], testconfig['genesis_delta'])
     _ = add_multi_clients(init_session, cspec, clients_num)
 
     hits = []
